# api/swagger.yml
...
swagger: "2.0"
info:
  version: 0.0.0
  title: Miasma
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/apps/{appName}/config:
    get:
      summary: get an app's current config
      operationId: getRunConfig
      parameters:
        - $ref: "#/parameters/appName"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/RunConfig"
        404:
          description: Not Found
          schema:
            type: string
        default:
          $ref: "#/responses/unknown"
    put:
      summary: update an app's config
      operationId: updateRunConfig
      parameters:
        - $ref: "#/parameters/appName"
        - in: body
          name: newRunConfig
          schema:
            $ref: "#/definitions/InputRunConfig"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/RunConfig"
        400:
          description: Bad Request
          schema:
            type: string
        404:
          description: Not Found
          schema:
            type: string
        default:
          $ref: "#/responses/unknown"
...
parameters:
  appName:
    name: appName
    in: path
    required: true
    type: string
    description: App name from the `-a|--app` flag
...
responses:
  unknown:
    description: Unknown Error
    schema:
      type: string
definitions:
...
  RunConfigVolume:
    type: object
    properties:
      Source:
        description: The volume name or directory on the host that the data is stored in
        type: string
      Target:
        description: The path inside the container that the data is served from
        type: string
      # Type:
      #   description:
      #     The type of binding. Should be either "bind", "volume", "tmpfs", or "npipe".
      #   type: string
      #   x-nullable: true
  RunConfig:
    type: object
    description: All the properties that define how the application runs in docker
    required:
      - appId
      - imageDigest
    properties:
      appId:
        description: The ID of the app the run config is for
        type: string
        format: uuid4
        x-go-custom-tag: 'gorm:"primaryKey"'
        x-nullable: false
        x-go-custom-tag: 'gorm:"primaryKey"'
      imageDigest:
        type: string
        description: |
          The currently running image digest (hash). Used internally when running applications
          instead of the tag because the when a new image is pushed, the tag stays the same but the
          digest changes
        x-nullable: false
      targetPorts:
        description:
          The ports that the app is listening to inside the container. If no target ports
          are specified, then the container should respect the `PORT` env var.
        type: array
        items:
          type: integer
          format: uint32
        uniqueItems: true
      publishedPorts:
        description:
          The ports that you access the app through in the swarm. This field can, and should
          be left empty. Miasma automatically manages assigning published ports between 3001-4999.
          If you need to specify a port, make sure it's outside that range or the port has not been
          taken. Plugins have set ports starting with 4000, so avoid 4000-4020 if you want to add a
          plugin at a later date.
          If these ports are ever cleared, the app will continue using the same ports it was
          published to before, so that the ports don't change unnecessarily. If you removed it to
          clear a port for another app/plugin, make sure to restart the app and a new, random port
          will be allocated for the app, freeing the old port
        type: array
        items:
          type: integer
          format: uint32
        uniqueItems: true
      placement:
        description:
          The placement constraints specifying which nodes the app will be ran on. Any valid
          value for the
          [`--constraint` flag](https://docs.docker.com/engine/swarm/services/#placement-constraints)
          is valid item in this list
        type: array
        items:
          type: string
        uniqueItems: true
      volumes:
        description: volume bindings for the app
        type: array
        items:
          $ref: "#/definitions/RunConfigVolume"
      networks:
        description:
          A list of other apps that the service communicates with using their service name and
          docker's internal DNS. Services don't have to be two way; only the service that accesses
          the other needs the other network added
        type: array
        items:
          type: string
        uniqueItems: true
      command:
        type: array
        items:
          type: string
  InputRunConfig:
    type: object
    description: All the properties that define how the application runs in docker
    properties:
      targetPorts:
        type: array
        items:
          type: integer
          format: uint32
        uniqueItems: true
      publishedPorts:
        type: array
        items:
          type: integer
          format: uint32
        uniqueItems: true
      placement:
        type: array
        items:
          type: string
        uniqueItems: true
      volumes:
        description: volume bindings for the app
        type: array
        items:
          $ref: "#/definitions/RunConfigVolume"
      networks:
        type: array
        items:
          type: string
        uniqueItems: true
      command:
        type: array
        items:
          type: string
...